design upd

+++ SPO_RKOT/client/src/components/Table/TableItem.js
@@ -1,31 +1,34 @@
-import React, { useContext, useState } from 'react';
+import React, {useContext, useState} from 'react';
 import ButtonUI from '../ButtonUI';
 import Row from 'react-bootstrap/esm/Row';
 import Col from 'react-bootstrap/esm/Col';
 import ExternalTableStore from '../../store/ExternalTableStore';
-import { Context } from '../..';
+import {Context} from '../..';
 
-const TableItem = ({ openTable, index, getIndex, dataTable, getDataList, ...props }) => {
+const TableItem = ({openTable, index, getIndex, dataTable, getDataList, ...props}) => {
 
     const [confirm, setConfirm] = useState(false);
 
-    const { externalTable } = useContext(Context)
+    const {externalTable} = useContext(Context)
     const openView = () => {
         openTable(true);
         getIndex(index, dataTable.id);
     }
 
     const deleteDataTable = async () => {
-        await externalTable.delete(dataTable.id)
-        getDataList()
+        if (window.confirm("Вы точно хотите удалить отчет?")) {
+            await externalTable.delete(dataTable.id)
+            getDataList()
+        }
         setConfirm(false)
     }
+
 
     return (
         <>
             <Row className='table_item_row'>
                 <Col className='table_column d-flex align-items-center justify-content-center'
-                    style={{ textAlign: "center", marginLeft: '7px' }}>{dataTable.district}</Col>
+                     style={{textAlign: "center", marginLeft: '7px'}}>{dataTable.district}</Col>
                 <Col
                     className='table_column_second d-flex align-items-center justify-content-center'>{dataTable.place}</Col>
                 <Col
